requirements.txt
django~=5.1
pip~=24.1.2
attrs~=23.2.0
djangorestframework ~= 3.15.2

